Added mutation for sending team invite emails
