+++ src/middleware/index.js
@@ -1,5 +1,3 @@
 const { shopifyValidator } = require('./shopifyValidator');
-const { sendEmail } = require('./email');
 
 exports.shopifyValidator = shopifyValidator;
-exports.sendEmail = sendEmail;
